Contour cache params now use correct data types

+++ include/vapor/ContourRenderer.h
@@ -69,13 +69,13 @@ class RENDER_API ContourRenderer : public Renderer {
     struct {
         string varName;
         string heightVarName;
-        int ts;
+        size_t ts;
         int level;
         int lod;
         bool useSingleColor;
         float constantColor[3];
-        float lineThickness;
-        float opacity;
+        double lineThickness;
+        double opacity;
         vector<double> boxMin, boxMax;
         vector<double> contourValues;
         vector<float> contourColors;
